refactor: extract buildCleanupGroupIds to cleanup-groups.ts and fix log level
